break down getUpdateTransactionPatch functions
