+++ apps/ledger-live-desktop/src/renderer/families/evm/EditTransaction/steps/StepFees.tsx
@@ -2,7 +2,7 @@ import { getMainAccount } from "@ledgerhq/live-common/account/index";
 import {
   getFormattedFeeFields,
   getMinFees,
-} from "@ledgerhq/live-common/families/evm/getUpdateTransactionPatch";
+} from "@ledgerhq/live-common/families/evm/editTransaction/index";
 import React, { Fragment, memo } from "react";
 import { Trans } from "react-i18next";
 import { useSelector } from "react-redux";
