+++ libs/ledger-live-common/src/families/evm/editTransaction/hasMinimumFunds.ts
@@ -1,0 +1,46 @@
+import { getEstimatedFees } from "@ledgerhq/coin-evm/logic";
+import type { Transaction } from "@ledgerhq/coin-evm/types/index";
+import { getEnv } from "@ledgerhq/live-env";
+import type { Account, AccountLike } from "@ledgerhq/types-live";
+import { BigNumber } from "bignumber.js";
+
+export const hasMinimumFundsToCancel = ({
+  mainAccount,
+  transactionToUpdate,
+}: {
+  mainAccount: Account;
+  transactionToUpdate: Transaction;
+}): boolean => {
+  const isEip1559 = transactionToUpdate.type === 2;
+
+  const factor: number = isEip1559
+    ? getEnv("EVM_REPLACE_TX_EIP1559_MAXFEE_FACTOR")
+    : getEnv("EVM_REPLACE_TX_LEGACY_GASPRICE_FACTOR");
+
+  const feeValue = getEstimatedFees(transactionToUpdate);
+  return mainAccount.balance.gt(feeValue.times(factor).integerValue(BigNumber.ROUND_CEIL));
+};
+
+export const hasMinimumFundsToSpeedUp = ({
+  account,
+  mainAccount,
+  transactionToUpdate,
+}: {
+  account: AccountLike;
+  mainAccount: Account;
+  transactionToUpdate: Transaction;
+}): boolean => {
+  const isEip1559 = transactionToUpdate.type === 2;
+
+  const factor: number = isEip1559
+    ? getEnv("EVM_REPLACE_TX_EIP1559_MAXFEE_FACTOR")
+    : getEnv("EVM_REPLACE_TX_LEGACY_GASPRICE_FACTOR");
+
+  const feeValue = getEstimatedFees(transactionToUpdate);
+  return mainAccount.balance.gt(
+    feeValue
+      .times(factor)
+      .plus(account.type === "Account" ? transactionToUpdate.amount : 0)
+      .integerValue(BigNumber.ROUND_CEIL),
+  );
+};
