+++ libs/ledger-live-common/src/families/evm/editTransaction/types.ts
@@ -1,0 +1,1 @@
+export type EditType = "cancel" | "speedup";
